image_engine/requirements.txt
face_recognition==1.3.0
numpy==1.23.0
pika==1.3.2
Pillow==10.0.1
python-dotenv==1.0.1
